requirements.txt
# --- web ---
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# --- validation / settings ---
pydantic>=2.0.0
pydantic-settings>=2.0.0

# --- logging ---
loguru>=0.7.0

# --- Bedrock (LangChain); pulls boto3 transitively ---
langchain-aws>=0.2.0
boto3>=1.34.0
